Consolidate duplicate mock fixtures and improve test isolation
- Move `mock_tf_env` fixture to `tests/conftest.py` for suite-wide reuse.
- Update `reset_globals` in `conftest.py` to reset `_model_cache` and `_async_openai_client`.
- Refactor `test_archive_scan.py`, `test_url_scan.py`, `test_extra_snippets.py`, `test_markdown_scan.py`, and `test_universal_parsing.py` to use the shared fixture.
- Ensure proper test isolation and remove redundant local definitions.
